DataStore Initialization
The DataStore is now initialized in the `MainActivity` using a new `initDataStore` function. This function takes an Android `Context` and stores it for later use in creating the DataStore instance. The `createDataStore` function now relies on this initialized context to retrieve the DataStore, throwing an exception if it hasn't been initialized.